+++ CHANGELOG.md
@@ -3,6 +3,9 @@
  - Fixed bug downloading files without extension
  - Replaced deprecated File.exists by File.exist function to avoid errors on newer Ruby versions
  - Added user-agent option (Thanks gmkbenjamin)
+ - AMSI bypass refactor
+ - ETW bypasss added
+ - Fixed Donut-Loader feature
 
 ### 3.5
  - Improvements on powershell functions loading